Clean and move resources folder to main source set
