+++ ktor-template-tests/build.gradle
@@ -26,6 +26,7 @@ dependencies {
 
 test {
     useJUnitPlatform()
+    forkEvery = 1
 }
 
 tasks.withType(Test) {
