+++ ktor-template-tests/src/main/resources/db/migration/V1.1__Initial_database_setup.sql
