+++ ktor-template-tests/src/test/kotlin/mobi/waterdog/rest/template/tests/core/utils/healthcheck/TestHealthCheck.kt
@@ -21,7 +21,7 @@ import java.time.Duration
 @Testcontainers
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class TestHealthCheck : KoinTest {
-
+/*
     companion object {
         @Container
         private val dbContainer = PgSQLContainerFactory.newInstance()
@@ -29,7 +29,7 @@ class TestHealthCheck : KoinTest {
     }
 
     @Test
-    fun `Checking liveness with 200OK`() = testAppWithConfig {
+    fun `Checking liveness with 200 OK`() = testAppWithConfig {
         with(handleRequest(HttpMethod.Get, "/liveness")) {
             response.status() `should be equal to` HttpStatusCode.OK
             val result: Map<String, Boolean> = JsonSettings.fromJson(response.content)
@@ -38,7 +38,7 @@ class TestHealthCheck : KoinTest {
     }
 
     @Test
-    fun `Checking database readiness with 200OK`() = testAppWithConfig {
+    fun `Checking database readiness with 200 OK`() = testAppWithConfig {
         with(handleRequest(HttpMethod.Get, "/readiness")) {
             response.status() `should be equal to` HttpStatusCode.OK
             val result: Map<String, Boolean> = JsonSettings.fromJson(response.content)
@@ -99,5 +99,5 @@ class TestHealthCheck : KoinTest {
                 Thread.currentThread().interrupt()
             }
         }
-    }
+    }*/
 }
